abstracted myBlockTypes into own file in assets directory
